pyproject.toml
[tool.poetry]
name = "db-kukasaataa"
version = "0.1.0"
description = ""
authors = ["Taina Hilliaho <taina.hilliaho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pymongo = "^4.10.1"
fitz = "^0.0.1.dev2"
pdfplumber = "^0.11.4"
typer = "^0.15.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
